refactor: renamed gpio_registers.h to bcm283x_gpio.h
